requirements.txt
pypdf
langchain
torch
accelerate
bitsandbytes
ctransformers
sentence_transformers
faiss_cpu
chainlit
huggingface_hub

asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
config==0.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.2
Flask-Cors==4.0.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
idna==3.6
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mistune==3.0.2
mongoengine==0.27.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
pymongo==4.6.2
python-dateutil==2.8.2
python-dotenv==1.0.1

pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.1
WTForms==3.1.2
yarg==0.1.9
